chy/raglangchain/pyproject.toml
[project]
name = "raglangchain"
version = "0.1.0"
description = "langchain-Project"
readme = "README.md"
requires-python = ">=3.11,<3.12"

dependencies = [
	# 언어 및 자연어 처리 관련 패키지
	"langchain-core",
	"langchain-experimental",
	"langchain-community",
	"langchain-openai",
	"langchain-anthropic",
	"langchain-text-splitters",
	"langchain-elasticsearch",
	"langchain-chroma",
	"langchain-cohere",
	"langchain-milvus",
	"langchain-google-genai",
	"langchain-huggingface",
	"langchain-azure-ai",
	"langchain-teddynote",
	"langchainhub",
	"langgraph",
	"langsmith",
	"huggingface-hub",
	"openai",
	"deepl",
	"kiwipiepy",
	"konlpy",
	# 데이터 처리 및 분석 관련 패키지
	"pandas",
	"rank-bm25",
	# 데이터베이스 및 캐시 관련 패키지
	"redis",
	"chromadb",
	# PDF 및 파일 처리 관련 패키지
	"pymupdf",
	"pypdf",
	"pdfplumber",
	"pdfminer-six",
	"pymupdf4llm",
	# 시각화 및 사용자 인터페이스 관련 패키지
	"matplotlib",
	"streamlit",
	"jupyter",
	"notebook",
	# 딥러닝 및 머신러닝 관련 패키지
	"torch",
	"torchvision",
	"faiss-cpu",
	"open-clip-torch",
	# 기타 유틸리티 및 필수 패키지
	"python-dotenv",
	"pydantic",
	"lxml",
	"pillow",
	"lark",
	"ragas",
	"unstructured[all-docs]",
	"arxiv",
	"tiktoken",
	"tenacity",
	"pymilvus",
	"google-search-results",
	"protobuf",
	"sqlalchemy",
	"llama-index-core",
	"llama-parse",
	"llama-index-readers-file",
	"flashrank",
	"docx2txt",
	"nest-asyncio",
	"rapidocr-onnxruntime",
	"seaborn",
	"grandalf",
	"rouge-score",
	"langchain-ollama",
	"mypy",
	"pinecone",
	"wikipedia",
	"scikit-learn",
	"hydra-core",
	"pytz",
	"anthropic",
	"gradio>=5.43.1",
	"omegaconf>=2.3.0",
	"requests>=2.32.5",
	"sentence-transformers>=5.1.0",
	"ipykernel>=6.30.1",
	"einops>=0.8.1",
]

